+++ src/test/java/com/github/jlangch/venice/impl/functions/IpcFunctionsTest.java
@@ -51,6 +51,25 @@ public class IpcFunctionsTest {
     }
 
     @Test
+    public void test_send_receive_async() {
+        final Venice venice = new Venice();
+
+        final String script =
+                "(do                                                     \n" +
+                "  (defn echo-handler [m] m)                             \n" +
+                "                                                        \n" +
+                "  (try-with [server (ipc/server 33333 echo-handler)     \n" +
+                "             client (ipc/client \"localhost\" 33333)]   \n" +
+                "    (->> (ipc/plain-text-message \"test\" \"hello\")    \n" +
+                "         (ipc/send-async client)                        \n" +
+                "         (deref)                                        \n" +
+                "         (ipc/message->map)                             \n" +
+                "         (:text ))))                                    ";
+
+        assertEquals("hello", venice.eval(script));
+    }
+
+    @Test
     public void test_send_receive_no_compress() {
         final Venice venice = new Venice();