+++ ChangeLog.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 
 - Simplified 'for' macro sequence expression partitioning into groups
 
+- Enhanced IPC api with async functions returning futures
+
 
 
 ## [1.12.60] - 2025-11-14